+++ src/datumaro/plugins/transforms.py
@@ -36,6 +36,7 @@ from datumaro.components.annotation import (
     Polygon,
     PolyLine,
     RleMask,
+    Tabular,
     TabularCategories,
 )
 from datumaro.components.cli_plugin import CliPlugin
@@ -59,10 +60,11 @@ from datumaro.components.errors import (
     UndefinedAttribute,
     UndefinedLabel,
 )
-from datumaro.components.media import Image
+from datumaro.components.media import Image, TableRow
 from datumaro.components.transformer import ItemTransform, Transform
 from datumaro.util import NOTSET, filter_dict, parse_json_file, parse_str_enum_value, take_by
 from datumaro.util.annotation_util import find_group_leader, find_instances
+from datumaro.util.tabular_util import emoji_pattern
 
 
 class CropCoveredSegments(ItemTransform, CliPlugin):
@@ -1391,6 +1393,9 @@ class Correct(Transform, CliPlugin):
         self._far_from_mean_caption = defaultdict(list)
         self._far_from_mean_value = {}
 
+        self._far_from_mean_caption = defaultdict(list)
+        self._far_from_mean_value = {}
+
         self._analyze_reports(report=self._reports)
 
         self._table = None
@@ -1526,20 +1531,6 @@ class Correct(Transform, CliPlugin):
             return re.sub(r"\s+", " ", text).strip()
 
         def remove_emojis(text):
-            emoji_pattern = re.compile(
-                "|".join(
-                    [
-                        "[\U0001F600-\U0001F64F]",  # emoticons
-                        "[\U0001F300-\U0001F5FF]",  # symbols & pictographs
-                        "[\U0001F680-\U0001F6FF]",  # transport & map symbols
-                        "[\U0001F1E0-\U0001F1FF]",  # flags (iOS)
-                        "[\u2600-\u26FF]",  # Miscellaneous Symbols
-                        "[\u2700-\u27BF]",  # Dingbats
-                        "[\U0001F900-\U0001F9FF]",  # Supplemental Symbols and Pictographs
-                    ]
-                ),
-                flags=re.UNICODE,
-            )
             return emoji_pattern.sub(r"", text)
 
         col_redun_dict = {}
@@ -1641,6 +1632,28 @@ class Correct(Transform, CliPlugin):
             )
         return annotations
 
+    def cap_far_from_mean(self, annotations, far_from_mean_captions):
+        for ann in annotations:
+            if ann.type != AnnotationType.caption:
+                continue
+
+            for col in far_from_mean_captions:
+                if not ann.caption.startswith(col):
+                    continue
+
+                value_str = ann.caption[len(col) + 1 :]
+                value = self.caption_type[col](value_str)
+
+                lower_bound, upper_bound = self._far_from_mean_value[col]
+                capped_value = max(min(value, upper_bound), lower_bound)
+
+                new_ann = Caption(f"{col}:{capped_value}")
+                annotations.remove(ann)
+                annotations.append(new_ann)
+                break
+
+        return annotations
+
     def cap_outliers(self, annotations, outliers):
         for ann in annotations:
             if ann.type != AnnotationType.caption:
@@ -1663,26 +1676,21 @@ class Correct(Transform, CliPlugin):
 
         return annotations
 
-    def cap_far_from_mean(self, annotations, far_from_mean_captions):
+    def find_outliers(self, annotations, outliers):
         for ann in annotations:
-            if ann.type != AnnotationType.caption:
-                continue
-
-            for col in far_from_mean_captions:
-                if not ann.caption.startswith(col):
-                    continue
-
-                value_str = ann.caption[len(col) + 1 :]
-                value = self.caption_type[col](value_str)
-
-                lower_bound, upper_bound = self._far_from_mean_value[col]
-                capped_value = max(min(value, upper_bound), lower_bound)
-
-                new_ann = Caption(f"{col}:{capped_value}")
-                annotations.remove(ann)
-                annotations.append(new_ann)
-                break
-
+            for col in outliers:
+                if ann.type == AnnotationType.caption and ann.caption[: len(col)] == col:
+                    value = self._extractor._tabular_cat_types[col](ann.caption[len(col) + 1 :])
+                    # Cap outliers
+                    lower_bound, upper_bound = self._outlier_value[col]
+                    cap_outlier_val = np.where(
+                        value > upper_bound,
+                        upper_bound,
+                        np.where(value < lower_bound, lower_bound, value),
+                    ).item()
+                    new_ann = Caption(f"{col}:{cap_outlier_val}")
+                    annotations.remove(ann)
+                    annotations.append(new_ann)
         return annotations
 
     def __iter__(self):
@@ -1837,3 +1845,134 @@ class AstypeAnnotations(ItemTransform):
         ]
 
         return self.wrap_item(item, annotations=annotations)
+
+
+class Clean(ItemTransform):
+    """ """
+
+    def __init__(
+        self,
+        extractor: IDataset,
+    ):
+        super().__init__(extractor)
+
+        self._outlier_value = {}
+        self._missing_value = {}
+
+    @staticmethod
+    def remove_unneccessary_char(text):
+        if pd.isna(text):
+            return text
+        try:
+            from nltk.corpus import stopwords
+
+            stop_words = set(stopwords.words("english"))  # TODO
+        except LookupError:
+            import nltk
+
+            nltk.download("stopwords")
+            stop_words = set(stopwords.words("english"))  # TODO
+
+        text = re.sub(r"<.*?>", "", text)  # Remove HTML tags
+        text = re.sub(r"http\S+|www\S+|https\S+", "", text, flags=re.MULTILINE)  # remove URLs
+        text = emoji_pattern.sub(r"", text)  # Remove emojis
+        text = text.lower()  # Convert to lowercase
+        text = re.sub(r"[^A-Za-z\s]+", "", text)  # Remove special characters and punctuation
+        text = re.sub(r"\s+", " ", text).strip()  # Remove extra whitespaces
+        text = " ".join(
+            [word for word in text.split() if word not in stop_words]
+        )  # Remove stopwords
+        return text
+
+    def check_outlier(self, table, numeric_cols):
+        for col in numeric_cols:
+            col_data = table[col].dropna()
+
+            Q1 = np.quantile(col_data, 0.25)
+            Q3 = np.quantile(col_data, 0.75)
+            IQR = Q3 - Q1
+            lower_bound = Q1 - 1.5 * IQR
+            upper_bound = Q3 + 1.5 * IQR
+
+            if table[col].dtype == int:
+                lower_bound = self.find_closest_value(col_data[col], lower_bound)
+                upper_bound = self.find_closest_value(col_data[col], upper_bound)
+            self._outlier_value[col] = (lower_bound, upper_bound)
+
+    def check_missing_value(self, table, float_cols, countable_cols):
+        from scipy.stats import skew
+
+        for col in table.columns:
+            col_data = table[col].dropna()
+            if col in float_cols:
+                skewness = skew(col_data)
+                if abs(skewness) > 1:  # TODO
+                    self._missing_value[col] = col_data.median()
+                else:
+                    self._missing_value[col] = col_data.mean()
+            elif col in countable_cols:
+                self._missing_value[col] = col_data.mode().iloc[0]
+
+    @staticmethod
+    def find_closest_value(series, target_value):
+        abs_diff = np.abs(series - target_value)
+        closest_index = abs_diff.idxmin()
+        return series.iloc[closest_index]
+
+    def cap_outliers(self, table):
+        lower, upper = self._outlier_value[table.name]
+        val = table.iloc[0]
+        if (val < lower) | (val > upper):
+            capped_value = max(min(val, upper), lower)
+            return table.replace(val, capped_value)
+        return val
+
+    def fill_missing_value(self, series):
+        return series.fillna(self._missing_value[series.name])
+
+    def refine_tabular_media(self, item):
+        media = item.media
+        df = pd.DataFrame(media.data(), index=[media.index])
+        str_cols = [col for col in media.data().keys() if item.media.table.dtype(col) is str]
+        float_cols = [col for col in media.data().keys() if item.media.table.dtype(col) is float]
+        int_cols = [col for col in media.data().keys() if item.media.table.dtype(col) is int]
+        countable_cols = [
+            col
+            for col in media.data().keys()
+            if isinstance(item.media.table.dtype(col), CategoricalDtype)
+            or item.media.table.dtype(col) is int
+        ]
+
+        df[str_cols] = df[str_cols].applymap(lambda x: self.remove_unneccessary_char(x))
+
+        if not (self._outlier_value):
+            self.check_outlier(media.table.data[float_cols + int_cols], float_cols + int_cols)
+        df[float_cols + int_cols] = df[float_cols + int_cols].apply(lambda x: self.cap_outliers(x))
+
+        if not (self._missing_value):
+            self.check_missing_value(
+                media.table.data[float_cols + countable_cols], float_cols, countable_cols
+            )
+        df[float_cols + countable_cols] = df[float_cols + countable_cols].apply(
+            lambda x: self.fill_missing_value(x)
+        )
+
+        return TableRow.from_data(
+            df.iloc[0].to_dict(), table=item.media.table, index=item.media.index
+        )
+
+    def transform_item(self, item):
+        if not isinstance(item.media, TableRow):
+            raise DatumaroError(
+                "Item %s: TableRow info is required for this " "transform" % (item.id,)
+            )
+
+        refined_media = self.refine_tabular_media(item) if item.media.has_data else None
+        annotation_values = {
+            key: refined_media.data[key] for key in item.annotations[0].values.keys()
+        }  # only for tabular
+        refined_annotations = [
+            ann.wrap(values=annotation_values) if isinstance(ann, Tabular) else ann
+            for ann in item.annotations
+        ]
+        return self.wrap_item(item, media=refined_media, annotations=refined_annotations)